Add tests/ with some basic reset tests
Start to add tests using pytest.

This lets us test resets on PCIe cards and Galaxy 6U systems.

For now I only added a basic reset test and a reset stress test for each
type of system.

Hopefully we can build this out with more tests and get this into GitHub
CI. We will need to figure out how we want to do recovery on runners.
Until then we can run this locally.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>